requirements.txt
APScheduler==3.11.0
asgiref==3.10.0
charset-normalizer==3.4.3
cssselect2==0.8.0
Django==5.2.7
django-cors-headers==4.9.0
django-debug-toolbar==6.0.0
django-filer==3.3.2
django-filter==25.2
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-mptt==0.18.0
django-polymorphic==4.1.0
django-unfold==0.68.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
easy-thumbnails==2.10.1
freetype-py==2.5.1
lxml==6.0.2
pillow==11.3.0
pycairo==1.28.0
PyJWT==2.10.1
python-decouple==3.8
reportlab==4.4.4
rlPyCairo==0.4.0
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
webencodings==0.5.1
